requirements.txt
fastapi
uvicorn[standard]
cryptography
python-dotenv
requests
pydantic
sqlalchemy
pymysql
alembic
pytest
httpx
